% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R, R/default.R, R/glm.R, R/lm.R,
%   R/svyglm.R, R/svymnlogit.R, R/svyolr.R, R/wickr.R
\name{ascribe.clm}
\alias{ascribe.clm}
\alias{ascribe.default}
\alias{ascribe.glm}
\alias{ascribe.lm}
\alias{ascribe.svyglm}
\alias{ascribe.svymnlogit}
\alias{ascribe.svyolr}
\alias{ascribe}
\title{Model statistics.}
\usage{
\method{ascribe}{clm}(x, ...)

\method{ascribe}{default}(x, ...)

\method{ascribe}{glm}(x, ...)

\method{ascribe}{lm}(x, vcov. = NULL,
  vcov_name = deparse(substitute(vcov.)), ...)

\method{ascribe}{svyglm}(x, ...)

\method{ascribe}{svymnlogit}(x, ...)

\method{ascribe}{svyolr}(x, ...)

ascribe(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default implementation which returns \code{NA}.}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}

\item{vcov_name}{character. A name to call the covariance matrix. Default \code{deparse(substitute(vcov))}.}
}
\value{
a list.
}
\description{
The default method returns \code{as.list(\link[broom]{glance}(x))}.
}
\section{Methods (by class)}{
\itemize{
\item \code{clm}: residual df, AIC, BIC, link, and n

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: OLS

\item \code{svyglm}: survey GLM

\item \code{svymnlogit}: only the link function ("logit")

\item \code{svyolr}: residual df, deviance, link, and n
}}

