% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R, R/default.R, R/wickr.R
\name{delevel.amce}
\alias{delevel.amce}
\alias{delevel.default}
\alias{delevel}
\title{Finish the work started by \code{\link{unlevel}}.}
\usage{
\method{delevel}{amce}(x, y, z, ...)

\method{delevel}{default}(x, y, z, ...)

delevel(x, y, z, ...)
}
\arguments{
\item{x}{A model object. Typically ignored (used only for method dispatch).}

\item{y}{Output of \code{\link{tidify}}.}

\item{z}{Output of \code{\link{unlevel}}.}

\item{...}{Ignored in the default implementation.}
}
\value{
a tidy presentation of results with separate term and level columns.
}
\description{
The default method will join \code{z} to \code{y} and return the relevant columns.
}
\section{Methods (by class)}{
\itemize{
\item \code{amce}: returns y; this step is handled in the \code{\link{tidify}} method

\item \code{default}: The method for otherwise unhandled objects.
}}

