% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg.R, R/default.R, R/wickr.R
\name{encomp.betareg}
\alias{encomp.betareg}
\alias{encomp.default}
\alias{encomp}
\title{Separate model component from name of term in MLEs where scale and shape are modeled.}
\usage{
\method{encomp}{betareg}(x, y, ...)

\method{encomp}{default}(x, y, ...)

encomp(x, y, ...)
}
\arguments{
\item{x}{A model object. Dispatch is on this argument.}

\item{y}{A data.frame put out by \code{\link{tidify}}.}

\item{...}{Ignored in the default method. Probably ignored in most other methods, too.}
}
\value{
A \code{data.frame}.
}
\description{
The default method is intended for models of the mean only and simply returns the input.
}
\section{Methods (by class)}{
\itemize{
\item \code{betareg}: non-trivial method for \code{betareg} objects

\item \code{default}: Do nothing to otherwise unhandled objects
}}

