% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R, R/betareg.R, R/clm.R, R/crch.R,
%   R/default.R, R/glm.R, R/lm.R, R/simpool.R, R/svymnlogit.R, R/svyolr.R,
%   R/wickr.R
\name{tidify.amce}
\alias{tidify.amce}
\alias{tidify.betareg}
\alias{tidify.clm}
\alias{tidify.crch}
\alias{tidify.default}
\alias{tidify.glm}
\alias{tidify.lm}
\alias{tidify.simpool}
\alias{tidify.svymnlogit}
\alias{tidify.svyolr}
\alias{tidify}
\title{Tabulate model results tidily.}
\usage{
\method{tidify}{amce}(x, margins = FALSE, .sdiff = FALSE, ...)

\method{tidify}{betareg}(x, margins = FALSE, data = if
  (is.null(x$model)) {     margins::find_data(x) } else {     x$model },
  variables = NULL, ...)

\method{tidify}{clm}(x, margins = FALSE, ...)

\method{tidify}{crch}(x, margins = FALSE, ...)

\method{tidify}{default}(x, margins = FALSE, ...)

\method{tidify}{glm}(x, margins = FALSE, ...)

\method{tidify}{lm}(x, margins = FALSE, vcov. = stats::vcov,
  variables = NULL, ...)

\method{tidify}{simpool}(x, ...)

\method{tidify}{svymnlogit}(x, margins = FALSE, ...)

\method{tidify}{svyolr}(x, margins = margins, ...)

tidify(x, margins = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{margins}{logical. Serve up marginal effects from package \code{\link[margins]{margins-package}}? Default \code{FALSE} (output regression coefficients).}

\item{.sdiff}{logical. Second differences? Default \code{FALSE}. Currently ignored, though.}

\item{...}{other arguments to \code{margins::\link[margins]{margins}} or \code{lmtest::coeftest}.}

\item{data}{data.frame. Where is the data located? Defaults to data within the model object.}

\item{variables}{character. Which terms to get effects for. Default \code{NULL} (\[everyone.gif\]).}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}
}
\value{
A \code{data.frame}.
}
\description{
The default method returns \code{broom::\link[broom]{tidy}(lmtest::\link[lmtest]{coeftest}(x, ...))}.
}
\section{Methods (by class)}{
\itemize{
\item \code{amce}: AMCE objects from package cjoint

\item \code{betareg}: \code{margins} plays strangely with \code{betareg} objects; this method works around those quirks

\item \code{clm}: uses coeftest

\item \code{crch}: uses coeftest

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: uses \code{lmtest::coeftest} with chosen covariance matrix.

\item \code{simpool}: uses coeftest

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: reaches into \code{survey} to get the right \code{vcov} for use in \code{lmtest::coeftest}
}}

