% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R, R/svymnlogit.R, R/wickr.R
\name{unlevel.default}
\alias{unlevel.default}
\alias{unlevel.svymnlogit}
\alias{unlevel}
\title{Replace concatenated term and level with separate columns.}
\usage{
\method{unlevel}{default}(x, ...)

\method{unlevel}{svymnlogit}(x, ...)

unlevel(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default implementation.}
}
\value{
a data.frame.
}
\description{
The default method will unpack \code{x$xlevels}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: The method for otherwise unhandled objects.

\item \code{svymnlogit}: svymnlogit
}}

