% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R, R/betareg.R, R/clm.R, R/crch.R,
%   R/default.R, R/glm.R, R/lm.R, R/simpool.R, R/svyglm.R, R/svymnlogit.R,
%   R/svyolr.R, R/wickr.R
\name{ascribe.amce}
\alias{ascribe.amce}
\alias{ascribe.betareg}
\alias{ascribe.clm}
\alias{ascribe.crch}
\alias{ascribe.default}
\alias{ascribe.glm}
\alias{ascribe.lm}
\alias{ascribe.simpool}
\alias{ascribe.svyglm}
\alias{ascribe.svymnlogit}
\alias{ascribe.svyolr}
\alias{ascribe}
\title{Model statistics.}
\usage{
\method{ascribe}{amce}(x, ...)

\method{ascribe}{betareg}(x, ...)

\method{ascribe}{clm}(x, ...)

\method{ascribe}{crch}(x, ...)

\method{ascribe}{default}(x, ...)

\method{ascribe}{glm}(x, ...)

\method{ascribe}{lm}(x, vcov. = NULL,
  vcov_name = deparse(substitute(vcov.)), ...)

\method{ascribe}{simpool}(x, ...)

\method{ascribe}{svyglm}(x, ...)

\method{ascribe}{svymnlogit}(x, ...)

\method{ascribe}{svyolr}(x, ...)

ascribe(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default implementation which returns \code{NA}.}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}

\item{vcov_name}{character. A name to call the covariance matrix. Default \code{deparse(substitute(vcov.))}.}
}
\value{
a list.
}
\description{
The default method returns \code{as.list(\link[broom]{glance}(x))}.
}
\section{Methods (by class)}{
\itemize{
\item \code{amce}: the only

\item \code{betareg}: residual df, AIC, BIC, link, and n

\item \code{clm}: residual df, AIC, BIC, link, and n

\item \code{crch}: AIC, BIC, n, and link

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: OLS

\item \code{simpool}: whatever's there

\item \code{svyglm}: survey GLM

\item \code{svymnlogit}: only the link function ("logit")

\item \code{svyolr}: residual df, deviance, link, and n
}}

