% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R, R/default.R, R/glm.R, R/lm.R,
%   R/svymnlogit.R, R/svyolr.R, R/wickr.R
\name{ascribe.clm}
\alias{ascribe.clm}
\alias{ascribe.default}
\alias{ascribe.glm}
\alias{ascribe.lm}
\alias{ascribe.svymnlogit}
\alias{ascribe.svyolr}
\alias{ascribe}
\title{Model statistics.}
\usage{
\method{ascribe}{clm}(x, ...)

\method{ascribe}{default}(x, ...)

\method{ascribe}{glm}(x, ...)

\method{ascribe}{lm}(x, vcov. = NULL, ...)

\method{ascribe}{svymnlogit}(x, ...)

\method{ascribe}{svyolr}(x, ...)

ascribe(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default implementation which returns \code{NA}.}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}
}
\value{
a list.
}
\description{
The default method returns \code{as.list(\link[broom]{glance}(x))}.
}
\section{Methods (by class)}{
\itemize{
\item \code{clm}: clm

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: OLS

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: svyolr
}}

