% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R, R/crch.R, R/default.R, R/glm.R, R/lm.R,
%   R/svymnlogit.R, R/svyolr.R, R/wickr.R
\name{kind.clm}
\alias{kind.clm}
\alias{kind.crch}
\alias{kind.default}
\alias{kind.glm}
\alias{kind.lm}
\alias{kind.svymnlogit}
\alias{kind.svyolr}
\alias{kind}
\title{Name of model.}
\usage{
\method{kind}{clm}(x, ...)

\method{kind}{crch}(x, ...)

\method{kind}{default}(x, ...)

\method{kind}{glm}(x, ...)

\method{kind}{lm}(x, ...)

\method{kind}{svymnlogit}(x, ...)

\method{kind}{svyolr}(x, ...)

kind(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default implementation which returns \code{NA}.}
}
\value{
a character vector.
}
\description{
The default method returns \code{class(x)}.
}
\section{Methods (by class)}{
\itemize{
\item \code{clm}: "ordinal"

\item \code{crch}: "censored"

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: "OLS"

\item \code{svymnlogit}: "multinomial"

\item \code{svyolr}: "ordinal"
}}

