% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R, R/default.R, R/gilez.R, R/glm.R,
%   R/list.R, R/lm.R, R/sumer.R, R/svymnlogit.R, R/svyolr.R, R/wickr.R
\name{sumer.clm}
\alias{sumer.clm}
\alias{sumer.default}
\alias{sumer.gilez}
\alias{sumer.glm}
\alias{sumer.list}
\alias{sumer.lm}
\alias{sumer.sumer}
\alias{sumer.svymnlogit}
\alias{sumer.svyolr}
\alias{sumer}
\title{Tidily summarize a model, with metadata.}
\usage{
\method{sumer}{clm}(x, ...)

\method{sumer}{default}(x, ...)

\method{sumer}{gilez}(x, ...)

\method{sumer}{glm}(x, ...)

\method{sumer}{list}(x, .depth = 0, .names = NULL, ...)

\method{sumer}{lm}(x, vcov. = NULL, ...)

\method{sumer}{sumer}(x, ...)

\method{sumer}{svymnlogit}(x, ...)

\method{sumer}{svyolr}(x, ...)

sumer(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default method.}

\item{.depth}{integer. Levels from top of nested list of objects. Default \code{0}.}

\item{.names}{character. Vector of names to represent levels, from the top downward. Default \code{NULL}.}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}
}
\value{
A \code{data.frame} with an added attribute or three.
}
\description{
The default method mostly wraps functions from \code{\link[broom]{broom}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{clm}: clm

\item \code{default}: The method for otherwise unhandled objects.

\item \code{gilez}: simulation results from \code{gilez}, which are always tall

\item \code{glm}: GLM

\item \code{list}: Unrolls lists of objects to be tidily summarized.

\item \code{lm}: OLS

\item \code{sumer}: Idempotent method.

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: svyolr
}}

