% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R, R/default.R, R/glm.R, R/lm.R,
%   R/svymnlogit.R, R/svyolr.R, R/wickr.R
\name{tidify.clm}
\alias{tidify.clm}
\alias{tidify.default}
\alias{tidify.glm}
\alias{tidify.lm}
\alias{tidify.svymnlogit}
\alias{tidify.svyolr}
\alias{tidify}
\title{Tabulate model results tidily.}
\usage{
\method{tidify}{clm}(x, ...)

\method{tidify}{default}(x, ...)

\method{tidify}{glm}(x, ...)

\method{tidify}{lm}(x, vcov. = NULL, ...)

\method{tidify}{svymnlogit}(x, ...)

\method{tidify}{svyolr}(x, ...)

tidify(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default method. Not passed to \code{tidy} in the default method.}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}
}
\value{
A \code{data.frame}.
}
\description{
The default method returns \code{\link[broom]{tidy}}, which will work for all the most common models.
}
\section{Methods (by class)}{
\itemize{
\item \code{clm}: clm

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: OLS

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: svyolr
}}

