% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R, R/betareg.R, R/clm.R, R/crch.R,
%   R/default.R, R/glm.R, R/lm.R, R/svymnlogit.R, R/svyolr.R, R/wickr.R
\name{tidify.amce}
\alias{tidify.amce}
\alias{tidify.betareg}
\alias{tidify.clm}
\alias{tidify.crch}
\alias{tidify.default}
\alias{tidify.glm}
\alias{tidify.lm}
\alias{tidify.svymnlogit}
\alias{tidify.svyolr}
\alias{tidify}
\title{Tabulate model results tidily.}
\usage{
\method{tidify}{amce}(x, .sdiff = FALSE, ...)

\method{tidify}{betareg}(x, ...)

\method{tidify}{clm}(x, ...)

\method{tidify}{crch}(x, ...)

\method{tidify}{default}(x, ...)

\method{tidify}{glm}(x, ...)

\method{tidify}{lm}(x, vcov. = NULL, ...)

\method{tidify}{svymnlogit}(x, ...)

\method{tidify}{svyolr}(x, ...)

tidify(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{.sdiff}{logical. Second differences? Default \code{FALSE}. Currently ignored, though.}

\item{...}{Ignored in the default method. Not passed to \code{tidy} in the default method.}

\item{vcov.}{a specification of the covariance matrix. See \code{\link[lmtest]{coeftest}} for details.}
}
\value{
A \code{data.frame}.
}
\description{
The default method returns \code{\link[broom]{tidy}}, which will work for all the most common models.
}
\section{Methods (by class)}{
\itemize{
\item \code{amce}: AMCE objects from package cjoint

\item \code{betareg}: uses coeftest

\item \code{clm}: uses coeftest

\item \code{crch}: uses coeftest

\item \code{default}: The method for otherwise unhandled objects.

\item \code{glm}: GLM

\item \code{lm}: uses \code{lmtest::coeftest} with chosen covariance matrix.

\item \code{svymnlogit}: svymnlogit

\item \code{svyolr}: reaches into \code{survey} to get the right \code{vcov} for use in \code{lmtest::coeftest}
}}

