% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R, R/betareg.R, R/crch.R, R/default.R,
%   R/simpool.R, R/svymnlogit.R, R/wickr.R
\name{unlevel.amce}
\alias{unlevel.amce}
\alias{unlevel.betareg}
\alias{unlevel.crch}
\alias{unlevel.default}
\alias{unlevel.simpool}
\alias{unlevel.svymnlogit}
\alias{unlevel}
\title{Replace concatenated term and level with separate columns.}
\usage{
\method{unlevel}{amce}(x, ...)

\method{unlevel}{betareg}(x, ...)

\method{unlevel}{crch}(x, ...)

\method{unlevel}{default}(x, ...)

\method{unlevel}{simpool}(x, ...)

\method{unlevel}{svymnlogit}(x, ...)

unlevel(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Ignored in the default implementation.}
}
\value{
a data.frame.
}
\description{
The default method will unpack \code{x$xlevels}.
}
\section{Methods (by class)}{
\itemize{
\item \code{amce}: returns NULL; this step is handled in the \code{\link{tidify}} method

\item \code{betareg}: \code{levels$full} replaces \code{xlevels}

\item \code{crch}: \code{levels$full} replaces \code{xlevels}

\item \code{default}: The method for otherwise unhandled objects.

\item \code{simpool}: OK then

\item \code{svymnlogit}: svymnlogit
}}

